ui: extract statements table component
Release note: None
